requirements.txt
# Core ML libraries
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.18.0
datasets>=2.0.0
sentencepiece>=0.1.96

# NLP tools
nltk>=3.6.7
spacy>=3.2.0
sacremoses>=0.0.53
stanza>=1.3.0

# Arabic NLP specific
pyarabic>=0.6.14
camel-tools>=1.3.0

# Evaluation metrics
sacrebleu>=2.0.0
rouge-score>=0.1.2
bert-score>=0.3.11

# API and web
flask>=2.0.1
fastapi>=0.75.0
uvicorn>=0.17.6
pydantic>=1.9.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
pandas>=1.4.2
numpy>=1.22.3
matplotlib>=3.5.1 